requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
openai
anthropic
dataforseo-client
mcp